pyproject.toml
[tool.poetry]
name = "brandquad"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.1"
python-dotenv = "^1.0.1"
djangorestframework = "^3.15.2"
psycopg2 = "^2.9.9"
psycopg2-binary = "^2.9.9"
flake8 = "^7.1.1"
pillow = "^10.4.0"
markdown = "^3.7"
django-filter = "^24.3"
drf-yasg = "^1.21.7"
setuptools = "^75.1.0"
django-cors-headers = "^4.4.0"
coverage = "^7.6.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
